CSS cleanup
Two changes in this patch:

(1) Two repeating CSS chunks were moved to their own files:

- Yahoo's "reset" styles were moved to '_reset_css.lua'.
- The code prettifier's styles were moved to '_code_css.lua'.

(The underscore at the filename's start is to show that it's not an
independent file but intended to be embedded in another.)

(2) Everything in "pre.example" that's already in "pre" was removed
(leaving just the font-size).

CAVEAT: the two new files need to be added to the 'rockspec' file.