Fix end-of-block comments not highlighting
Originally, comments in code blocks were highlighted by looking for
their starting token at the beginning and a newline at the end.

This PR adds two new comment patterns that look for the end-of-string
instead of a newline, so that comments on the last line of a block are
still highlighted.