Update makefile
Removed `DESTDIR` which did not appear to be defined, moved` ldoc.lua` to the `ldoc` directory to avoid encumbering the share directory, added proper retrieval of Lua version to get proper `LUA_SHAREDIR` directory. Additionally, the file should be renamed as `Makefile`.

My apologies if I just missed something and this change proposal is irrelevant.