fix: Return original t in `find_module` if `search_for_token` fails (fixes #258)
If `search_for_token` does not find a `module` token, t is nil, and t is
returned instead of the original t.

This is an issue, because it causes a branch in `parse.lua` to
erroneously be executed, which exits the entire parsing process with the
warning `contains no items`, even if items are left (because the
original t was scrubbed).